Add `depends_on` to server
The default ordering will attempt to delete the firewall policy and
group and fail because the server is still running.

So this frequently failed and left the policy. A second destroy was
needed to clear up (after the server was gone).

This dependency priorities the destruction of the server (the dependent)
on the policy (the dependency) allowing more reliable clean ups.