Fixes for Terraform updates
The provider and version have changed since this example was first
created. Required providers and their versions are declared outside of
the provider code now.

The Brightbox provider has now moved from a `hashicorp` prefix to the
registry under `brightbox/brightbox`.

A `sensitive` setting is required for the DB password output to avoid an
error during planning/applying so the field is protected.